+++ src/test/java/projet/projetstage02/controller/RootControllerTest.java
@@ -44,7 +44,6 @@ import projet.projetstage02.model.Token;
 import projet.projetstage02.service.AuthService;
 import projet.projetstage02.service.CompanyService;
 import projet.projetstage02.service.GestionnaireService;
-import projet.projetstage02.service.StudentService;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
@@ -65,22 +64,19 @@ import static projet.projetstage02.utils.TimeUtil.currentTimestamp;
 public class RootControllerTest {
 
     MockMvc mockMvc;
-    MockMvc studentMockMvc;
     MockMvc companyMockMvc;
     MockMvc gestionnaireMockMvc;
 
     @InjectMocks
     RootController rootController;
-    @InjectMocks
-    StudentRootController studentRootController;
+
     @InjectMocks
     CompanyRootController companyRootController;
     @InjectMocks
     GestionnaireRootController gestionnaireRootController;
 
 
-    @Mock
-    StudentService studentService;
+
     @Mock
     CompanyService companyService;
     @Mock
@@ -88,18 +84,15 @@ public class RootControllerTest {
     @Mock
     GestionnaireService gestionnaireService;
 
-    JacksonTester<StudentInDTO> jsonStudentDTO;
     JacksonTester<LoginDTO> jsonLoginDTO;
     JacksonTester<CompanyDTO> jsonCompanyDTO;
     JacksonTester<GestionnaireDTO> jsonGestionnaireDTO;
     JacksonTester<OffreInDTO> jsonOffreDTO;
     JacksonTester<TokenDTO> jsonTokenDTO;
     JacksonTester<CvRefusalDTO> jsonCvRefusalDTO;
-    JacksonTester<PdfDTO> jsonPdfDTO;
     JacksonTester<SignatureInDTO> jsonSignatureDTO;
     JacksonTester<MillieuStageEvaluationInDTO> jsonEvalInDTO;
     JacksonTester<InterviewOutDTO> jsonInterviewOutDTO;
-    JacksonTester<InterviewSelectInDTO> jsonInterviewSelectDTO;
 
     StudentInDTO bart;
     StudentOutDTO bartOut;
@@ -202,7 +195,6 @@ public class RootControllerTest {
         JacksonTester.initFields(this, new ObjectMapper());
 
         mockMvc = MockMvcBuilders.standaloneSetup(rootController).build();
-        studentMockMvc = MockMvcBuilders.standaloneSetup(studentRootController).build();
         companyMockMvc = MockMvcBuilders.standaloneSetup(companyRootController).build();
         gestionnaireMockMvc = MockMvcBuilders.standaloneSetup(gestionnaireRootController).build();
 
@@ -330,57 +322,6 @@ public class RootControllerTest {
                 .studentId(bart.getId())
                 .selectedDate("2022-11-29T12:30")
                 .build();
-    }
-
-    @Test
-    void testCreateStudentHappyDay() throws Exception {
-        when(studentService.isStudentInvalid(anyString())).thenReturn(false);
-        when(studentService.saveStudent(any())).thenReturn(1L);
-        studentMockMvc.perform(post("/createStudent")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonStudentDTO.write(bart).getJson()))
-
-                .andExpect(status().isCreated());
-    }
-
-    @Test
-    void testCreateStudentConflict() throws Exception {
-        when(studentService.isStudentInvalid(anyString())).thenReturn(true);
-
-        studentMockMvc.perform(post("/createStudent")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonStudentDTO.write(bart).getJson()))
-
-                .andExpect(status().isConflict());
-    }
-
-    @Test
-    void testCreateStudentBadRequest() throws Exception {
-        studentMockMvc.perform(post("/createStudent")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonStudentDTO.write(new StudentInDTO()).getJson()))
-
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void testCreateStudentExistingEmailDeleteOld() throws Exception {
-        when(studentService.isStudentInvalid(anyString())).thenReturn(false);
-        studentMockMvc.perform(post("/createStudent")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonStudentDTO.write(bart).getJson()))
-
-                .andExpect(status().isCreated());
-    }
-
-    @Test
-    void testCreateStudentExistingEmailDeleteOldNotFound() throws Exception {
-        when(studentService.isStudentInvalid(anyString())).thenThrow(new NonExistentEntityException());
-        studentMockMvc.perform(post("/createStudent")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonStudentDTO.write(bart).getJson()))
-
-                .andExpect(status().isInternalServerError());
     }
 
     @Test
@@ -524,37 +465,6 @@ public class RootControllerTest {
                 .andExpect(status().isForbidden());
     }
 
-    @Test
-    void testConfirmStudentEmailHappyDay() throws Exception {
-        when(studentService.getStudentById(1L)).thenReturn(bartOut);
-        when(studentService.saveStudent(any())).thenReturn(1L);
-
-        studentMockMvc.perform(
-                        put("/confirmEmail/student/{id}", 1))
-
-                .andExpect(status().isCreated());
-    }
-
-    @Test
-    void testConfirmStudentEmailNotFound() throws Exception {
-        when(studentService.getStudentById(1L)).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(
-                        put("/confirmEmail/student/{id}", 1))
-
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testConfirmStudentEmailExpired() throws Exception {
-        bartOut.setInscriptionTimestamp(Timestamp.valueOf(LocalDateTime.now().minusMonths(1)).getTime());
-        when(studentService.getStudentById(1L)).thenReturn(bartOut);
-
-        studentMockMvc.perform(
-                        put("/confirmEmail/student/{id}", 1))
-
-                .andExpect(status().isBadRequest());
-    }
 
     @Test
     void testConfirmCompanyEmailHappyDay() throws Exception {
@@ -634,19 +544,6 @@ public class RootControllerTest {
                 .andExpect(jsonPath("$.token", is(token.getToken())));
     }
 
-    @Test
-    void testLoginStudentHappyDay() throws Exception {
-        bartOut.setEmailConfirmed(true);
-        when(authService.getToken(token.getToken(), STUDENT)).thenReturn(Token.builder().userId(1L).build());
-        when(studentService.getStudentById(1L)).thenReturn(bartOut);
-
-        studentMockMvc.perform(put("/student")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.firstName", is("Bart")));
-    }
 
     @Test
     void testLoginStudentNotEmailConfirmed() throws Exception {
@@ -660,19 +557,6 @@ public class RootControllerTest {
 
                 .andExpect(status().isForbidden());
         verify(authService, times(1)).loginIfValid(any(), any());
-    }
-
-    @Test
-    void testLoginStudentNotFound() throws Exception {
-        when(authService.getToken(token.getToken(), STUDENT)).thenReturn(Token.builder().userId(1).build());
-        when(studentService.getStudentById(anyLong()))
-                .thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/student")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-
-                .andExpect(status().isNotFound());
     }
 
     @Test
@@ -1088,41 +972,7 @@ public class RootControllerTest {
                 .andExpect(status().isForbidden());
     }
 
-    @Test
-    void testUploadCurriculumVitaeSuccess() throws Exception {
-        when(authService.getToken(any(), any())).thenReturn(Token.builder().userId(1).build());
-        bartOut.setCvToValidate(byteToString(bartCV.getPdf()));
-        when(studentService.uploadCurriculumVitae(any())).thenReturn(bartOut);
 
-        studentMockMvc.perform(put("/uploadStudentCV")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonPdfDTO.write(bartCV).getJson()))
-
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.firstName", is("Bart")))
-                .andExpect(jsonPath("$.cvToValidate", is("[]")));
-    }
-
-    @Test
-    void testUploadCurriculumVitaeNotFound() throws Exception {
-        when(authService.getToken(any(), any())).thenReturn(Token.builder().userId(1).build());
-        doThrow(new NonExistentEntityException())
-                .when(studentService).uploadCurriculumVitae(any());
-
-        studentMockMvc.perform(put("/uploadStudentCV")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(jsonPdfDTO.write(bartCV).getJson()));
-    }
-
-    @Test
-    void testUploadCurriculumVitaeInvalidToken() throws Exception {
-        when(authService.getToken(any(), any())).thenThrow(new InvalidTokenException());
-
-        studentMockMvc.perform(put("/uploadStudentCV")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonPdfDTO.write(bartCV).getJson()))
-                .andExpect(status().isForbidden());
-    }
 
     @Test
     void testGetOfferPdfSuccess() throws Exception {
@@ -1287,144 +1137,6 @@ public class RootControllerTest {
     }
 
     @Test
-    void testGetOffersByStudentDepartmentSuccess() throws Exception {
-        when(authService.getToken(any(), any())).thenReturn(Token.builder().userId(1).build());
-        when(studentService.getOffersByStudentDepartment(anyLong())).thenReturn(List.of(duffOffreOut));
-
-        studentMockMvc.perform(put("/getOffers/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].nomDeCompagnie", is(duffOffre.getNomDeCompagnie())));
-    }
-
-    @Test
-    void testGetOffersByStudentDepartmentNotFound() throws Exception {
-        when(authService.getToken(any(), any())).thenReturn(Token.builder().userId(1).build());
-        when(studentService.getOffersByStudentDepartment(anyLong())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/getOffers/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testGetOffersByStudentDepartmentInvalidToken() throws Exception {
-        when(authService.getToken(any(), any())).thenThrow(new InvalidTokenException());
-        studentMockMvc.perform(put("/getOffers/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void testGetOfferStudentSuccess() throws Exception {
-        when(authService.getToken(any(), any())).thenReturn(Token.builder().userId(1).build());
-        when(studentService.getOfferPdfById(anyLong())).thenReturn(duffOfferOut);
-
-        studentMockMvc.perform(put("/getOfferStudent/{id}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.pdf", is(duffOfferOut.getPdf())));
-    }
-
-    @Test
-    void testGetOfferStudentNotFound() throws Exception {
-        when(authService.getToken(any(), any())).thenReturn(Token.builder().userId(1).build());
-        when(studentService.getOfferPdfById(anyLong())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/getOfferStudent/{id}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testGetOfferStudentInvalidToken() throws Exception {
-        when(authService.getToken(any(), any())).thenThrow(new InvalidTokenException());
-
-        studentMockMvc.perform(put("/getOfferStudent/{id}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void testCreatePostulationSuccess() throws Exception {
-        when(studentService.createPostulation(anyLong(), anyLong())).thenReturn(bartPostulation);
-        studentMockMvc.perform(put("/applyToOffer/{studentId}_{offerId}", 2, 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.fullName", is(bartPostulation.getFullName())))
-                .andExpect(jsonPath("$.company", is(bartPostulation.getCompany())));
-    }
-
-    @Test
-    void testCreatePostulationNotFound() throws Exception {
-        when(studentService.createPostulation(anyLong(), anyLong())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/applyToOffer/{studentId}_{offerId}", 2, 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testCreatePostulationConflict() throws Exception {
-        when(studentService.createPostulation(anyLong(), anyLong())).thenThrow(new AlreadyExistingPostulation());
-
-        studentMockMvc.perform(put("/applyToOffer/{studentId}_{offerId}", 2, 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isConflict());
-    }
-
-    @Test
-    void testCreatePostulationInvalidToken() throws Exception {
-        when(authService.getToken(any(), any())).thenThrow(new InvalidTokenException());
-
-        studentMockMvc.perform(put("/applyToOffer/{studentId}_{offerId}", 2, 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void testGetPostulsOfferIdSuccess() throws Exception {
-        when(studentService.getPostulsOfferId(anyLong())).thenReturn(bartApplys);
-
-        studentMockMvc.perform(put("/studentApplys/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.offersId.size()", is(1)));
-    }
-
-    @Test
-    void testGetPostulsOfferIdNotFound() throws Exception {
-        when(studentService.getPostulsOfferId(anyLong())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/studentApplys/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testGetPostulsOfferIdTokenInvalid() throws Exception {
-        when(authService.getToken(any(), any())).thenThrow(new InvalidTokenException());
-
-        studentMockMvc.perform(put("/studentApplys/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
     void testSaveStudentAcceptationHappyDay() throws Exception {
         when(companyService.saveStudentApplicationAccepted(anyLong(), anyLong())).thenReturn(applicationDTO);
 
@@ -1538,36 +1250,6 @@ public class RootControllerTest {
                 .andExpect(status().isForbidden());
     }
 
-    @Test
-    void testGetStatusValidationCVHappyDay() throws Exception {
-        when(studentService.getStudentCvStatus(anyLong())).thenReturn(cvStatusDTO);
-
-        studentMockMvc.perform(put("/getStatutValidationCV/{id}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.status", is("ACCEPTED")));
-    }
-
-    @Test
-    void testGetStatusValidationCVForbidden() throws Exception {
-        when(authService.getToken(any(), any())).thenThrow(new InvalidTokenException());
-
-        studentMockMvc.perform(put("/getStatutValidationCV/{id}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void testGetStatusValidationCVNotFound() throws Exception {
-        when(studentService.getStudentCvStatus(anyLong())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/getStatutValidationCV/{id}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isNotFound());
-    }
 
     @Test
     void testGetCompanyOffersHappyDay() throws Exception {
@@ -1708,36 +1390,6 @@ public class RootControllerTest {
                 .andExpect(status().isForbidden());
     }
 
-    @Test
-    void testStudentContractSignatureHappyDay() throws Exception {
-        when(studentService.addSignatureToContract(any())).thenReturn(stageContractOutDTO);
-
-        studentMockMvc.perform(put("/studentSignatureContract")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonSignatureDTO.write(signatureInDTO).getJson()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.studentSignature", is(signatureInDTO.getSignature())));
-    }
-
-    @Test
-    void testStudentContractSignatureNotFound() throws Exception {
-        when(studentService.addSignatureToContract(any())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/studentSignatureContract")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonSignatureDTO.write(signatureInDTO).getJson()))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testStudentContractSignatureConflict() throws Exception {
-        when(studentService.addSignatureToContract(any())).thenThrow(new InvalidOwnershipException());
-
-        studentMockMvc.perform(put("/studentSignatureContract")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonTokenDTO.write(token).getJson()))
-                .andExpect(status().isConflict());
-    }
 
     @Test
     void testGetUnvalidatedAcceptationsInvalidToken() throws Exception {
@@ -1920,88 +1572,4 @@ public class RootControllerTest {
                 .andExpect(status().isForbidden());
     }
 
-    @Test
-    void testStudentSelectDateHappyDay() throws Exception{
-        when(studentService.selectInterviewTime(any())).thenReturn(interviewOutDTO);
-
-        studentMockMvc.perform(put("/studentSelectDate")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    void testStudentSelectDateNotFound() throws Exception{
-        when(studentService.selectInterviewTime(any())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/studentSelectDate")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testStudentSelectDateForbidden() throws Exception{
-        when(studentService.selectInterviewTime(any())).thenThrow(new InvalidOwnershipException());
-
-        studentMockMvc.perform(put("/studentSelectDate")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void testStudentSelectDateBadRequest() throws Exception{
-        when(studentService.selectInterviewTime(any())).thenThrow(new InvalidDateFormatException());
-
-        studentMockMvc.perform(put("/studentSelectDate")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void testStudentSelectDateInvalidToken() throws Exception{
-        when(authService.getToken(any(), any())).thenThrow(new InvalidTokenException());
-
-        studentMockMvc.perform(put("/studentSelectDate")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void testGetStudentInterviewsHappyDay() throws Exception{
-        when(studentService.getInterviews(anyLong())).thenReturn(new ArrayList<>(){{
-            add(new InterviewOutDTO());
-            add(new InterviewOutDTO());
-            add(new InterviewOutDTO());
-        }});
-
-        studentMockMvc.perform(put("/getStudentInterviews/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.size()", is(3)));
-    }
-
-    @Test
-    void testGetStudentInterviewsNotFound() throws Exception{
-        when(studentService.getInterviews(anyLong())).thenThrow(new NonExistentEntityException());
-
-        studentMockMvc.perform(put("/getStudentInterviews/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void testGetStudentInterviewsInvalidToken() throws Exception{
-        when(authService.getToken(anyString(), any())).thenThrow(new InvalidTokenException());
-
-        studentMockMvc.perform(put("/getStudentInterviews/{studentId}", 1)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(jsonInterviewSelectDTO.write(interviewSelectInDTO).getJson()))
-                .andExpect(status().isForbidden());
-    }
 }
